Refactor desktop frontend structure
