+++ apps/desktop/src/App.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { invoke } from '@tauri-apps/api/core'
-  import { emitTo, listen } from '@tauri-apps/api/event'
+  import { listen } from '@tauri-apps/api/event'
   import { getCurrentWebview } from '@tauri-apps/api/webview'
   import {
     isPermissionGranted,
@@ -8,7 +8,6 @@
     sendNotification,
   } from '@tauri-apps/plugin-notification'
   import { revealItemInDir } from '@tauri-apps/plugin-opener'
-  import { ArrowUpRight, ChevronDown } from '@lucide/svelte'
   import { onMount, tick } from 'svelte'
 
   import rambelleArchived from './assets/rambelle-states/archived.png'
@@ -16,8 +15,10 @@
   import rambelleOrganizing from './assets/rambelle-states/organizing.png'
   import rambelleRecording from './assets/rambelle-states/recording.png'
   import AppTitlebar from './lib/AppTitlebar.svelte'
-  import RichFeedbackEditor from './lib/RichFeedbackEditor.svelte'
   import SettingsPanel from './lib/SettingsPanel.svelte'
+  import InboxPanel from './lib/workbench/InboxPanel.svelte'
+  import ResumePromptDialog from './lib/workbench/ResumePromptDialog.svelte'
+  import WorkspacePanel from './lib/workbench/WorkspacePanel.svelte'
   import type {
     AddAttachmentInput,
     AttachmentView,
@@ -31,12 +32,6 @@
     SaveDraftInput,
     SubmitFeedbackInput,
   } from './lib/feedback'
-  import { requestStatusLabel } from './lib/feedback'
-  import {
-    clipboardCaptureLabel,
-    eventBelongsToRamble,
-    type ClipboardCaptureEvent,
-  } from './lib/clipboardCapture'
   import type { HealthSnapshot } from './lib/generated/health'
   import {
     InboxNotificationTracker,
@@ -45,21 +40,17 @@
     type NotificationState,
   } from './lib/notifications'
   import { desktopPath } from './lib/nativePath'
-  import {
-    RAMBLE_CONSOLE_COMMAND_EVENT,
-    RAMBLE_CONSOLE_HIDE_EVENT,
-    RAMBLE_CONSOLE_READY_EVENT,
-    RAMBLE_CONSOLE_SHOW_EVENT,
-    RAMBLE_CONSOLE_STATE_EVENT,
-    type RambleConsoleCommand,
-    type RambleConsoleState,
-  } from './lib/rambleConsole'
-  import {
-    eventBelongsToVoiceSession,
-    stableTranscript,
-    type SpeechEvent,
-    type VoiceRambleSessionView,
-  } from './lib/speech'
+  import type {
+    AdapterPresentation,
+    FeedbackEditorHandle,
+    RamblePhase,
+    RambleSessionControllerHandle,
+    ResumePrompt,
+    SavePhase,
+    SettingsSection,
+    VoicePhase,
+  } from './lib/workbench/types'
+  import RambleSessionController from './lib/workbench/RambleSessionController.svelte'
   import type { ScreenCaptureReady } from './lib/screenCapture'
   import { t } from './lib/i18n'
   import { locale } from './lib/preferences'
@@ -69,25 +60,7 @@
     outcome: 'cancelled' | 'pinned'
   }
 
-  type SavePhase = 'idle' | 'unsaved' | 'saving' | 'saved' | 'error'
-  type RamblePhase = 'idle' | 'starting' | 'active' | 'paused' | 'stopping' | 'error'
-  type VoicePhase = 'idle' | 'starting' | 'listening' | 'processing' | 'stopping' | 'error'
   type CommandError = { code: string; message: string; retryable: boolean }
-  type SettingsSection = 'general' | 'mcp'
-  type ResumePrompt = {
-    request_id: string
-    host_id: string
-    host_label: string
-    title: string
-    body: string
-    resume_prompt: string
-    reason: 'completed' | 'cancelled'
-  }
-  type AdapterPresentation = {
-    id: string
-    label: string
-    icon_svg: string
-  }
 
   const RESUME_PROMPT_EVENT = 'rambledesk://resume-prompt'
 
@@ -113,8 +86,8 @@
   let attachmentMessage = ''
   let attachmentPreviews: Record<string, string> = {}
   let dragActive = false
-  let attachmentInput: HTMLInputElement
-  let richEditor: RichFeedbackEditor
+  let workspacePanel: FeedbackEditorHandle
+  let rambleController: RambleSessionControllerHandle
   let resumePrompt: ResumePrompt | null = null
   let resumeCopyState: 'idle' | 'copied' | 'failed' = 'idle'
   let notificationState: NotificationState = 'checking'
@@ -124,19 +97,13 @@
   let taskBriefOpen = true
   let mcpConfiguration = ''
   let voicePhase: VoicePhase = 'idle'
-  let voiceRequestId = ''
-  let voiceSessionId = ''
   let voiceDevice = ''
   let voicePartial = ''
-  let voiceMessage = ''
   let voiceLevel = 0
   let voiceChunkIndex = 0
   let ramblePhase: RamblePhase = 'idle'
   let rambleStartedOnce = false
-  let rambleContextId = ''
   let rambleMessage = ''
-  let clipboardCaptureCount = 0
-  let clipboardImageQueue: Promise<void> = Promise.resolve()
   let saveTimer: ReturnType<typeof setTimeout> | undefined
   let inboxTimer: ReturnType<typeof setInterval> | undefined
   let activeSave: Promise<boolean> | null = null
@@ -162,7 +129,7 @@
     voicePhase === 'processing' ||
     voicePhase === 'stopping'
   $: voiceCanStop =
-    voiceActive || (voicePhase === 'error' && voiceSessionId.length > 0)
+    voiceActive || voicePhase === 'error'
   $: rambleActive = ramblePhase === 'active'
   $: rambleEngaged = ramblePhase !== 'idle'
   $: rambelleStatusPortrait = feedbackResult
@@ -175,25 +142,6 @@
   $: rambleBusy = ramblePhase === 'starting' || ramblePhase === 'stopping'
   $: rambleCanStop = rambleActive || voiceCanStop
   $: rambleCanExit = rambleEngaged || voiceCanStop
-  $: if (rambleEngaged && workspace) {
-    const consoleState: RambleConsoleState = {
-      phase:
-        ramblePhase === 'active'
-          ? 'recording'
-          : ramblePhase === 'idle'
-            ? 'paused'
-            : ramblePhase,
-      projectName: workspace.request.project_name,
-      requestTitle: workspace.request.title,
-      recording: rambleActive,
-      busy: rambleBusy,
-      captureBusy: attachmentBusy,
-      voiceLevel,
-      partialTranscript: voicePartial,
-      message: rambleMessage,
-    }
-    void emitTo('ramble-console', RAMBLE_CONSOLE_STATE_EVENT, consoleState).catch(() => {})
-  }
 
   onMount(() => {
     if (!isTauri) {
@@ -206,13 +154,8 @@
     void refreshNotificationPermission()
     inboxTimer = setInterval(() => void refreshInbox(), 5_000)
     let dragUnlisten: (() => void) | undefined
-    let voiceUnlisten: (() => void) | undefined
-    let rambleShortcutUnlisten: (() => void) | undefined
-    let captureShortcutUnlisten: (() => void) | undefined
     let captureReadyUnlisten: (() => void) | undefined
     let captureFinishedUnlisten: (() => void) | undefined
-    let consoleCommandUnlisten: (() => void) | undefined
-    let consoleReadyUnlisten: (() => void) | undefined
     let resumePromptUnlisten: (() => void) | undefined
     void listen<ResumePrompt>(RESUME_PROMPT_EVENT, (event) => {
       resumePrompt = event.payload
@@ -231,35 +174,6 @@
       })
       .catch(() => {
         // Resume prompt still appears if submit path keeps the main window focused.
-      })
-    void listen<SpeechEvent>('voice-ramble-event', (event) => {
-      handleVoiceEvent(event.payload)
-    })
-      .then((unlisten) => {
-        voiceUnlisten = unlisten
-      })
-      .catch((cause) => {
-        voicePhase = 'error'
-        voiceMessage = tr('无法监听语音识别事件：{error}', { error: messageFrom(cause) })
-      })
-    void listen<string>('screen-capture-shortcut', () => {
-      if (rambleEngaged) void startScreenCapture()
-    })
-      .then((unlisten) => {
-        captureShortcutUnlisten = unlisten
-      })
-      .catch((cause) => {
-        attachmentMessage = tr('无法监听截图快捷键：{error}', { error: messageFrom(cause) })
-      })
-    void listen<string>('ramble-toggle-shortcut', () => {
-      void toggleRamble()
-    })
-      .then((unlisten) => {
-        rambleShortcutUnlisten = unlisten
-      })
-      .catch((cause) => {
-        ramblePhase = 'error'
-        rambleMessage = tr('无法监听 Ramble 快捷键：{error}', { error: messageFrom(cause) })
       })
     void listen<ScreenCaptureReady>('screen-capture-ready', (event) => {
       void importScreenCapture(event.payload)
@@ -281,16 +195,6 @@
       .catch(() => {
         // A failed cancellation listener does not affect capture or attachment storage.
       })
-    void listen<RambleConsoleCommand>(RAMBLE_CONSOLE_COMMAND_EVENT, (event) => {
-      void handleRambleConsoleCommand(event.payload)
-    }).then((unlisten) => {
-      consoleCommandUnlisten = unlisten
-    })
-    void listen(RAMBLE_CONSOLE_READY_EVENT, () => {
-      broadcastRambleConsoleState()
-    }).then((unlisten) => {
-      consoleReadyUnlisten = unlisten
-    })
     void getCurrentWebview()
       .onDragDropEvent((event) => {
         dragActive = event.payload.type === 'enter' || event.payload.type === 'over'
@@ -312,15 +216,9 @@
       if (saveTimer) clearTimeout(saveTimer)
       if (inboxTimer) clearInterval(inboxTimer)
       dragUnlisten?.()
-      voiceUnlisten?.()
-      rambleShortcutUnlisten?.()
-      captureShortcutUnlisten?.()
       captureReadyUnlisten?.()
       captureFinishedUnlisten?.()
-      consoleCommandUnlisten?.()
-      consoleReadyUnlisten?.()
       resumePromptUnlisten?.()
-      if (voiceCanStop) void invoke('stop_voice_ramble')
       window.removeEventListener('paste', handlePaste)
       releaseAttachmentPreviews()
     }
@@ -621,7 +519,7 @@
       }
       await refreshAttachmentPreviews(next)
       await tick()
-      const inserted = richEditor?.insertAttachments(
+      const inserted = workspacePanel?.insertAttachments(
         next.attachments.filter((item) => !existingIds.has(item.attachment_id)),
       )
       if (!inserted) {
@@ -654,7 +552,7 @@
       }
       await refreshAttachmentPreviews(next)
       await tick()
-      richEditor?.insertAttachments(
+      workspacePanel?.insertAttachments(
         next.attachments.filter((item) => !existingIds.has(item.attachment_id)),
       )
       await saveDraftNow()
@@ -715,7 +613,7 @@
       applyWorkspaceMutation(next)
       await refreshAttachmentPreviews(next)
       await tick()
-      const inserted = richEditor?.insertAttachments(
+      const inserted = workspacePanel?.insertAttachments(
         next.attachments.filter((item) => !existingIds.has(item.attachment_id)),
       )
       if (!inserted) {
@@ -736,7 +634,7 @@
 
   async function removeAttachment(attachment: AttachmentView) {
     if (!workspace || attachmentBusy) return
-    richEditor?.removeAttachmentReference(attachment.attachment_id)
+    workspacePanel?.removeAttachmentReference(attachment.attachment_id)
     if (!(await saveDraftNow())) return
     attachmentBusy = true
     attachmentMessage = ''
@@ -757,7 +655,7 @@
   }
 
   function insertExistingAttachment(attachment: AttachmentView) {
-    richEditor?.insertAttachments([attachment])
+    workspacePanel?.insertAttachments([attachment])
   }
 
   async function moveAttachment(index: number, offset: number) {
@@ -868,382 +766,24 @@
     }
   }
 
-  async function startRamble() {
-    if (
-      !workspace ||
-      rambleBusy ||
-      rambleEngaged ||
-      workspace.request.status === 'completed' ||
-      workspace.request.status === 'cancelled'
-    ) {
-      return
-    }
-    rambleStartedOnce = true
-    rambleContextId = crypto.randomUUID()
-    clipboardCaptureCount = 0
-    ramblePhase = 'starting'
-    rambleMessage = tr('正在打开 Ramble 操作台…')
-    void emitTo('ramble-console', RAMBLE_CONSOLE_SHOW_EVENT).catch((cause) => {
-      pageError = tr('无法打开 Ramble 操作台：{error}', { error: messageFrom(cause) })
-    })
-    await resumeRamble()
-  }
-
-  async function resumeRamble() {
-    if (!workspace || rambleBusy || rambleActive || !rambleContextId) return
-    const requestId = workspace.request.request_id
-    ramblePhase = 'starting'
-    rambleMessage = tr('正在启动麦克风与实时转写…')
-    const voiceStarted = await startVoiceRamble()
-    if (!voiceStarted || !voiceSessionId) {
-      ramblePhase = 'error'
-      rambleMessage = voiceMessage || tr('麦克风启动失败')
-      return
-    }
-
-    if (workspace?.request.request_id !== requestId) {
-      await invoke('stop_voice_ramble').catch(() => {})
-      resetVoiceUi()
-      await exitRamble()
-      return
-    }
-    ramblePhase = 'active'
-    rambleMessage = tr('Ramble 进行中 · 剪贴板仅在点击导入时读取')
-  }
-
-  async function stopRamble() {
-    if (!rambleCanStop || ramblePhase === 'stopping') return
-    ramblePhase = 'stopping'
-    rambleMessage = tr('正在收尾最后一段语音并暂停记录…')
-    let stopError = ''
-    if (voiceCanStop) {
-      const voiceStopped = await stopVoiceRamble()
-      if (!voiceStopped && !stopError) stopError = voiceMessage || tr('麦克风停止失败')
-    }
-    if (stopError) {
-      ramblePhase = 'error'
-      rambleMessage = stopError
-    } else {
-      ramblePhase = 'paused'
-      rambleMessage = tr('Ramble 已暂停；正文保留，截图和导入仍可使用')
-    }
-  }
-
   async function exitRamble() {
-    if (!rambleCanExit && !rambleStartedOnce) return
-    if (voiceCanStop) {
-      ramblePhase = 'stopping'
-      rambleMessage = tr('正在结束 Ramble…')
-      await stopVoiceRamble()
-    }
-    void emitTo('ramble-console', RAMBLE_CONSOLE_HIDE_EVENT).catch(() => {})
-    resetVoiceUi()
-    resetRambleUi()
+    await rambleController?.exitRamble()
   }
 
   async function toggleRamble() {
-    if (rambleBusy) return
-    if (rambleActive || voiceCanStop) await stopRamble()
-    else if (rambleEngaged) await resumeRamble()
-    else await startRamble()
+    await rambleController?.toggleRamble()
   }
 
   async function importClipboardNow() {
-    if (!workspace || !rambleEngaged || !rambleContextId || attachmentBusy) return
-    attachmentMessage = ''
-    try {
-      const event = await invoke<ClipboardCaptureEvent>('capture_clipboard_once', {
-        input: {
-          request_id: workspace.request.request_id,
-          ramble_session_id: rambleContextId,
-        },
-      })
-      handleClipboardCaptureEvent(event)
-    } catch (cause) {
-      attachmentMessage = tr('无法导入剪贴板：{error}', { error: messageFrom(cause) })
-    }
-  }
-
-  async function startVoiceRamble(): Promise<boolean> {
-    if (
-      !workspace ||
-      voiceActive ||
-      workspace.request.status === 'completed' ||
-      workspace.request.status === 'cancelled'
-    ) {
-      return false
-    }
-    voicePhase = 'starting'
-    voiceRequestId = workspace.request.request_id
-    voiceSessionId = ''
-    voiceDevice = ''
-    voicePartial = ''
-    voiceMessage = tr('正在加载本地模型并连接麦克风…')
-    voiceLevel = 0
-    try {
-      const session = await invoke<VoiceRambleSessionView>('start_voice_ramble', {
-        input: {
-          request_id: workspace.request.request_id,
-        },
-      })
-      voiceSessionId = session.session_id
-      if (voicePhase === 'starting') {
-        voicePhase = 'listening'
-        voiceMessage = tr('Sherpa 真流式识别 · 自然停顿后写入正文')
-      }
-    } catch (cause) {
-      voicePhase = 'error'
-      voiceMessage = messageFrom(cause)
-      return false
-    }
-    return true
-  }
-
-  async function stopVoiceRamble(): Promise<boolean> {
-    if (!voiceCanStop) return true
-    voicePhase = 'stopping'
-    voiceMessage = tr('正在完成最后一段识别…')
-    try {
-      await invoke('stop_voice_ramble')
-      for (let attempt = 0; attempt < 5 && voicePhase === 'stopping'; attempt += 1) {
-        await new Promise((resolve) => setTimeout(resolve, 20))
-      }
-      await tick()
-      if (voicePhase === 'stopping') {
-        voicePhase = 'idle'
-        voiceMessage = tr('录音已停止')
-      }
-    } catch (cause) {
-      voicePhase = 'error'
-      voiceMessage = messageFrom(cause)
-      return false
-    } finally {
-      voiceLevel = 0
-    }
-    return true
-  }
-
-  function handleClipboardCaptureEvent(event: ClipboardCaptureEvent) {
-    if (
-      !rambleEngaged ||
-      !workspace ||
-      !eventBelongsToRamble(
-        event,
-        workspace.request.request_id,
-        rambleContextId,
-      )
-    ) {
-      if (event.type === 'image') {
-        void invoke('discard_clipboard_capture_image', {
-          captureId: event.capture_id,
-        })
-      }
-      return
-    }
-
-    if (event.type === 'warning') {
-      rambleMessage = event.message
-      return
-    }
-    if (event.type === 'text') {
-      const inserted = richEditor?.appendClipboardCapture(
-        event.text,
-        clipboardCaptureLabel(event.captured_at_ms, event.truncated, $locale),
-      )
-      if (inserted) {
-        clipboardCaptureCount += 1
-        rambleMessage = tr('Ramble 进行中 · 已捕获 {count} 项剪贴板上下文', { count: clipboardCaptureCount })
-      }
-      return
-    }
-
-    clipboardImageQueue = clipboardImageQueue
-      .then(() => importClipboardImage(event))
-      .catch((cause) => {
-        attachmentMessage = tr('剪贴板图片写入失败：{error}', { error: messageFrom(cause) })
-      })
-  }
-
-  async function importClipboardImage(
-    event: Extract<ClipboardCaptureEvent, { type: 'image' }>,
-  ) {
-    const requestId = event.request_id
-    try {
-      for (let attempt = 0; attachmentBusy && attempt < 200; attempt += 1) {
-        await new Promise((resolve) => setTimeout(resolve, 50))
-      }
-      if (attachmentBusy) throw new Error(tr('附件通道正忙，请稍后重新复制图片'))
-      if (!workspace || workspace.request.request_id !== requestId) return
-      if (!(await saveDraftNow())) throw new Error(tr('当前草稿无法保存'))
-
-      attachmentBusy = true
-      const png = await invoke<ArrayBuffer>('read_clipboard_capture_image', {
-        captureId: event.capture_id,
-        requestId,
-        rambleSessionId: event.ramble_session_id,
-      })
-      const input: AddAttachmentInput = {
-        request_id: requestId,
-        file_name: event.file_name,
-        contents: Array.from(new Uint8Array(png)),
-        expected_revision: savedRevision,
-      }
-      const next = await invoke<FeedbackWorkspaceView>('add_feedback_attachment', { input })
-      if (workspace?.request.request_id !== requestId) return
-      const attachment = next.attachments.find(
-        (item) => !workspace?.attachments.some(
-          (existing) => existing.attachment_id === item.attachment_id,
-        ),
-      )
-      applyWorkspaceMutation(next)
-      await refreshAttachmentPreviews(next)
-      await tick()
-      if (
-        !attachment ||
-        !richEditor?.appendCapturedAttachment(
-          attachment,
-          clipboardCaptureLabel(event.captured_at_ms, false, $locale),
-        )
-      ) {
-        throw new Error(tr('图片附件已保存，但未能写入文档流'))
-      }
-      await saveDraftNow()
-      clipboardCaptureCount += 1
-      rambleMessage = tr('Ramble 进行中 · 已捕获 {count} 项剪贴板上下文', { count: clipboardCaptureCount })
-    } finally {
-      attachmentBusy = false
-      await invoke('discard_clipboard_capture_image', {
-        captureId: event.capture_id,
-      }).catch(() => {})
-    }
-  }
-
-  function handleVoiceEvent(event: SpeechEvent) {
-    const currentRequestId = workspace?.request.request_id ?? voiceRequestId
-    if (
-      !eventBelongsToVoiceSession(
-        event,
-        currentRequestId,
-        voiceSessionId,
-      )
-    ) {
-      return
-    }
-    voiceRequestId = event.request_id
-    voiceSessionId = event.session_id
-    switch (event.type) {
-      case 'started':
-        voicePhase = 'listening'
-        voiceDevice = event.input_device
-        voiceMessage = tr('正在录音 · {device}', { device: event.input_device })
-        break
-      case 'partial':
-        voicePartial = event.text
-        if (voicePhase !== 'stopping') voicePhase = 'listening'
-        break
-      case 'level':
-        voiceLevel = Math.min(1, Math.max(0, event.rms * 8))
-        if (voicePhase !== 'stopping') voicePhase = 'listening'
-        break
-      case 'processing':
-        voiceChunkIndex = event.chunk_index + 1
-        if (voicePhase !== 'stopping') voicePhase = 'processing'
-        voiceMessage = tr('正在识别第 {count} 段…', { count: event.chunk_index + 1 })
-        break
-      case 'stable': {
-        const transcript = stableTranscript(event)
-        if (transcript) richEditor?.appendTranscript(transcript)
-        voicePartial = ''
-        voiceChunkIndex = event.chunk_index + 1
-        if (voicePhase !== 'stopping') voicePhase = 'listening'
-        voiceMessage = tr('第 {count} 段已写入正文', { count: event.chunk_index + 1 })
-        break
-      }
-      case 'warning':
-        voiceMessage = event.message
-        break
-      case 'stopped':
-        voicePhase = 'idle'
-        voiceSessionId = ''
-        voiceLevel = 0
-        voicePartial = ''
-        voiceMessage = tr('录音已停止')
-        if (ramblePhase === 'active') {
-          ramblePhase = 'error'
-          rambleMessage = tr('麦克风意外停止，Ramble 已暂停')
-        }
-        break
-      case 'error':
-        voicePhase = 'error'
-        voiceLevel = 0
-        voicePartial = ''
-        voiceMessage = event.message
-        if (ramblePhase === 'active') {
-          ramblePhase = 'error'
-          rambleMessage = tr('麦克风错误，Ramble 已暂停：{error}', { error: event.message })
-        }
-        break
-    }
-  }
-
-  async function handleRambleConsoleCommand(command: RambleConsoleCommand) {
-    switch (command.type) {
-      case 'toggle-recording':
-        await toggleRamble()
-        break
-      case 'capture-screen':
-        await startScreenCapture()
-        break
-      case 'import-clipboard':
-        await importClipboardNow()
-        break
-      case 'import-files':
-        await importAttachmentPaths(command.paths)
-        break
-      case 'exit':
-        await exitRamble()
-        break
-    }
-  }
-
-  function broadcastRambleConsoleState() {
-    if (!rambleEngaged || !workspace) return
-    const state: RambleConsoleState = {
-      phase:
-        ramblePhase === 'active'
-          ? 'recording'
-          : ramblePhase === 'idle'
-            ? 'paused'
-            : ramblePhase,
-      projectName: workspace.request.project_name,
-      requestTitle: workspace.request.title,
-      recording: rambleActive,
-      busy: rambleBusy,
-      captureBusy: attachmentBusy,
-      voiceLevel,
-      partialTranscript: voicePartial,
-      message: rambleMessage,
-    }
-    void emitTo('ramble-console', RAMBLE_CONSOLE_STATE_EVENT, state).catch(() => {})
+    await rambleController?.importClipboardNow()
   }
 
   function resetVoiceUi() {
-    voicePhase = 'idle'
-    voiceRequestId = ''
-    voiceSessionId = ''
-    voiceDevice = ''
-    voicePartial = ''
-    voiceMessage = ''
-    voiceLevel = 0
-    voiceChunkIndex = 0
+    rambleController?.resetVoiceUi()
   }
 
   function resetRambleUi() {
-    ramblePhase = 'idle'
-    rambleStartedOnce = false
-    rambleContextId = ''
-    rambleMessage = ''
-    clipboardCaptureCount = 0
+    rambleController?.resetRambleUi()
   }
 
   function formatTime(value: string | null | undefined): string {
@@ -1272,6 +812,30 @@
 
 {#key $locale}
 <main class="shell">
+  <RambleSessionController
+    bind:this={rambleController}
+    {isTauri}
+    {workspace}
+    editor={workspacePanel}
+    bind:attachmentBusy
+    bind:attachmentMessage
+    {savedRevision}
+    bind:voicePhase
+    bind:voiceDevice
+    bind:voicePartial
+    bind:voiceLevel
+    bind:voiceChunkIndex
+    bind:ramblePhase
+    bind:rambleStartedOnce
+    bind:rambleMessage
+    onPageError={(message) => (pageError = message)}
+    onSaveDraftNow={saveDraftNow}
+    onApplyWorkspaceMutation={applyWorkspaceMutation}
+    onRefreshAttachmentPreviews={refreshAttachmentPreviews}
+    onStartScreenCapture={startScreenCapture}
+    onImportAttachmentPaths={importAttachmentPaths}
+  />
+
   <AppTitlebar
     projectName={workspace?.request.project_name ?? 'Vault Zero Archive'}
     connected={health?.status === 'ready'}
@@ -1285,408 +849,72 @@
   />
 
   <div class="workbench">
-    <aside class="inbox-panel">
-      <div class="panel-heading">
-        <div>
-          <p class="eyebrow">{inboxMode === 'open' ? 'INBOX' : 'HISTORY'}</p>
-          <h1>{inboxMode === 'open' ? tr('待反馈') : tr('历史记录')}</h1>
-        </div>
-        <button class="icon-button" aria-label={tr('刷新反馈请求')} onclick={refreshCurrentList}>↻</button>
-      </div>
+    <InboxPanel
+      {inboxMode}
+      {loadingInbox}
+      {loadingHistory}
+      requests={displayedRequests}
+      activeRequestId={workspace?.request.request_id ?? null}
+      {endpoint}
+      {adapterPresentation}
+      {formatTime}
+      onRefresh={refreshCurrentList}
+      onShowOpen={showOpenRequests}
+      onShowHistory={showHistory}
+      onOpenRequest={(requestId) => void openRequest(requestId)}
+      onOpenSettings={() => void openSettings('mcp')}
+    />
 
-      <div class="inbox-tabs" aria-label={tr('反馈列表范围')}>
-        <button class:active={inboxMode === 'open'} onclick={showOpenRequests}>{tr('待处理')}</button>
-        <button class:active={inboxMode === 'history'} onclick={showHistory}>{tr('全部历史')}</button>
-      </div>
-
-      {#if (inboxMode === 'open' && loadingInbox) || (inboxMode === 'history' && loadingHistory)}
-        <p class="empty-state">{tr('正在读取持久请求…')}</p>
-      {:else if displayedRequests.length === 0}
-        <div class="empty-state">
-          <strong>{inboxMode === 'open' ? tr('当前没有待处理请求') : tr('还没有反馈历史')}</strong>
-          <span>
-            {inboxMode === 'open'
-              ? tr('保持工作台开启，Agent 的新请求会出现在这里。')
-              : tr('创建过的请求会按最近更新时间显示在这里。')}
-          </span>
-        </div>
-      {:else}
-        <nav aria-label={inboxMode === 'open' ? tr('待反馈请求') : tr('反馈历史')}>
-          {#each displayedRequests as request (request.request_id)}
-            <button
-              class:active={workspace?.request.request_id === request.request_id}
-              class="request-card"
-              onclick={() => openRequest(request.request_id)}
-            >
-              <span class="request-meta">
-                <b>{request.project_name}</b>
-                <em>{requestStatusLabel(request.status, $locale)}</em>
-              </span>
-              {#if request.title.trim()}<strong>{request.title}</strong>{/if}
-              <small class="request-byline">
-                <span class="adapter-mark" aria-hidden="true">{@html adapterPresentation(request.agent).icon_svg}</span>
-                {adapterPresentation(request.agent).label} · {formatTime(request.updated_at)}
-              </small>
-            </button>
-          {/each}
-        </nav>
-      {/if}
-
-      <button
-        type="button"
-        class="connection-card"
-        aria-label={tr('打开 MCP 设置')}
-        onclick={() => void openSettings('mcp')}
-      >
-        <span>Local MCP</span>
-        <code>{endpoint}</code>
-        <ArrowUpRight size={15} strokeWidth={1.7} />
-      </button>
-    </aside>
-
-    <section class="workspace-panel">
-      {#if loadingWorkspace}
-        <div class="workspace-placeholder">{tr('正在打开反馈工作区…')}</div>
-      {:else if workspace}
-        <div class="workspace-stage">
-          <header class="workspace-heading">
-            <div class="workspace-heading-copy">
-              <div class="workspace-meta">
-                <span>{workspace.request.project_name}</span>
-                <span class="adapter-chip">
-                  <i class="adapter-mark" aria-hidden="true">{@html adapterPresentation(workspace.request.agent).icon_svg}</i>
-                  {adapterPresentation(workspace.request.agent).label}
-                </span>
-                <span class="status-chip">{requestStatusLabel(workspace.request.status, $locale)}</span>
-              </div>
-              {#if workspace.request.title.trim()}<h2>{workspace.request.title}</h2>{/if}
-              <p>Session · {workspace.request.session_id}</p>
-            </div>
-            <button class="secondary-button compact-button" onclick={reloadWorkspace}>{tr('重新载入')}</button>
-          </header>
-
-          <div class="workspace-columns">
-            <div class="document-column">
-              <section class:open={taskBriefOpen} class="task-sheet">
-                <button
-                  class="task-sheet-toggle"
-                  aria-expanded={taskBriefOpen}
-                  aria-label={taskBriefOpen ? tr('收起') : tr('展开')}
-                  title={taskBriefOpen ? tr('收起') : tr('展开')}
-                  onclick={() => (taskBriefOpen = !taskBriefOpen)}
-                >
-                  <span>
-                    <i>01</i>
-                    <strong>{tr('任务简报')}</strong>
-                    <em>{tr('{count} 个体验步骤', { count: workspace.actions.length })}</em>
-                  </span>
-                  <b class:open={taskBriefOpen} class="task-sheet-toggle-icon">
-                    <ChevronDown size={20} strokeWidth={1.9} />
-                  </b>
-                </button>
-
-                {#if taskBriefOpen}
-                  <div class="task-sheet-body">
-                    <section class="brief-section brief-summary-section">
-                      <p class="eyebrow">WHAT HAPPENED</p>
-                      <p class="brief-summary">{workspace.request.what_happened}</p>
-                    </section>
-
-                    <section class="brief-section brief-actions-section">
-                      <p class="eyebrow">WHAT TO TRY</p>
-                      <ol class="actions">
-                        {#each workspace.actions as action}
-                          <li>
-                            <span>{action.id}</span>
-                            <p>{action.instruction}</p>
-                          </li>
-                        {/each}
-                      </ol>
-                    </section>
-                  </div>
-                {/if}
-              </section>
-
-              <section class:drag-active={dragActive} class="editor-section">
-                <div class="editor-heading">
-                  <div>
-                    <p class="eyebrow">YOUR FEEDBACK</p>
-                    <h3>{tr('边体验，边记下来')}</h3>
-                  </div>
-                  <div class:failed={savePhase === 'error'} class="save-state" aria-live="polite">
-                    <span class="save-dot"></span>
-                    {#if savePhase === 'saving'}
-                      {tr('正在保存…')}
-                    {:else if savePhase === 'unsaved'}
-                      {tr('等待自动保存')}
-                    {:else if savePhase === 'error'}
-                      {tr('保存失败')}
-                    {:else}
-                      {tr('已保存')} · revision {savedRevision}
-                    {/if}
-                  </div>
-                </div>
-
-                <RichFeedbackEditor
-                  bind:this={richEditor}
-                  markdown={draftBody}
-                  previews={attachmentPreviews}
-                  disabled={workspace.request.status === 'completed' || workspace.request.status === 'cancelled'}
-                  onChange={updateDraft}
-                />
-
-                {#if attachmentMessage}
-                  <p class="inline-error">{attachmentMessage}</p>
-                {/if}
-
-                {#if saveMessage}
-                  <p class="inline-error">{saveMessage}。{tr('请重新载入后再试，当前文字仍保留在编辑器中。')}</p>
-                {/if}
-
-                <footer class="editor-footer">
-                  <span>{tr('{count} 字符', { count: draftBody.length.toLocaleString($locale) })}</span>
-                  <span>{tr('Markdown 文档流')}</span>
-                  <span>{formatTime(workspace.draft.updated_at)}</span>
-                </footer>
-              </section>
-            </div>
-
-            <aside class="command-rail" aria-label={tr('Ramble 操作台')}>
-              <section
-                class:active={rambleEngaged}
-                class:error={ramblePhase === 'error'}
-                class="ramble-console"
-              >
-                <div class="rail-heading">
-                  <div>
-                    <p class="eyebrow">RAMBLE</p>
-                    <strong>{rambleActive ? tr('正在记录') : rambleEngaged ? tr('Ramble 已暂停') : tr('记录待命')}</strong>
-                  </div>
-                  <span class="ramble-led"></span>
-                </div>
-
-                <button
-                  class:recording={rambleActive}
-                  class="ramble-primary"
-                  disabled={rambleBusy || workspace.request.status === 'completed' || workspace.request.status === 'cancelled'}
-                  onclick={toggleRamble}
-                  title={tr('全局快捷键 Ctrl + Shift + R')}
-                >
-                  <span>{rambleActive ? 'Ⅱ' : '●'}</span>
-                  {#if ramblePhase === 'starting'}
-                    {tr('正在启动…')}
-                  {:else if ramblePhase === 'stopping'}
-                    {tr('正在暂停…')}
-                  {:else if rambleActive}
-                    {tr('暂停 Ramble')}
-                  {:else if rambleStartedOnce}
-                    {tr('继续 Ramble')}
-                  {:else}
-                    {tr('开始 Ramble')}
-                  {/if}
-                </button>
-
-                {#if rambleEngaged}
-                  <button class="ramble-exit" disabled={rambleBusy} onclick={exitRamble}>
-                    {tr('退出 Ramble 操作台')}
-                  </button>
-                {/if}
-
-                <div class="voice-status">
-                  <div class="voice-title">
-                    <span class="voice-dot"></span>
-                    <strong>{voiceDevice || tr('默认麦克风')}</strong>
-                    {#if voiceChunkIndex > 0}<span>{tr('{count} 段', { count: voiceChunkIndex })}</span>{/if}
-                  </div>
-                  <span>{rambleMessage || tr('开始后可离开窗口继续操作，录音会实时写入正文。')}</span>
-                  {#if voicePartial}
-                    <em class="voice-partial">{tr('正在听：{text}', { text: voicePartial })}</em>
-                  {/if}
-                  <small>{tr('Sherpa X-ASR · 本地流式转写')}</small>
-                  <div class="voice-meter" aria-label={tr('麦克风音量')}>
-                    <span style={`width: ${voiceLevel * 100}%`}></span>
-                  </div>
-                </div>
-              </section>
-
-              <section class="tool-card">
-                <div class="rail-heading">
-                  <div>
-                    <p class="eyebrow">CAPTURE</p>
-                    <strong>{tr('添加上下文')}</strong>
-                  </div>
-                  <span>{workspace.attachments.length}</span>
-                </div>
-                <div class="tool-grid">
-                  <button
-                    disabled={!rambleEngaged || attachmentBusy || workspace.request.status === 'completed' || workspace.request.status === 'cancelled'}
-                    onclick={startScreenCapture}
-                    title="Ctrl + Shift + 1"
-                  >
-                    <span class="tool-icon">⌗</span>
-                    <strong>{tr('截图')}</strong>
-                    <small>{tr('区域捕获')}</small>
-                  </button>
-                  <button
-                    disabled={!rambleEngaged || attachmentBusy || workspace.request.status === 'completed' || workspace.request.status === 'cancelled'}
-                    onclick={importClipboardNow}
-                  >
-                    <span class="tool-icon">▣</span>
-                    <strong>{tr('剪贴板')}</strong>
-                    <small>{tr('显式导入')}</small>
-                  </button>
-                  <button
-                    disabled={attachmentBusy || workspace.request.status === 'completed' || workspace.request.status === 'cancelled'}
-                    onclick={() => attachmentInput.click()}
-                  >
-                    <span class="tool-icon">＋</span>
-                    <strong>{tr('文件')}</strong>
-                    <small>{tr('选择或拖入')}</small>
-                  </button>
-                </div>
-                <input
-                  bind:this={attachmentInput}
-                  class="visually-hidden"
-                  type="file"
-                  multiple
-                  onchange={handleFileSelection}
-                />
-                <p class="tool-hint">{tr('不会监听剪贴板；只有点击导入时才读取一次当前内容。')}</p>
-              </section>
-
-              <section class="attachments-card">
-                <div class="rail-heading">
-                  <div>
-                    <p class="eyebrow">ATTACHMENTS</p>
-                    <strong>{tr('文档素材')}</strong>
-                  </div>
-                  <span>{workspace.attachments.length}</span>
-                </div>
-                {#if workspace.attachments.length > 0}
-                  <div class="attachment-list" aria-label={tr('文档附件')}>
-                    {#each workspace.attachments as attachment, index (attachment.attachment_id)}
-                      <div class="attachment-row">
-                        <span class="attachment-dot"></span>
-                        <div>
-                          <strong>{attachment.file_name}</strong>
-                          <span>{(attachment.byte_size / 1024).toFixed(1)} KiB</span>
-                        </div>
-                        <div class="attachment-actions">
-                          <button
-                            aria-label={tr('插入正文 {name}', { name: attachment.file_name })}
-                            disabled={attachmentBusy || workspace.request.status === 'completed' || workspace.request.status === 'cancelled'}
-                            onclick={() => insertExistingAttachment(attachment)}
-                          >{tr('插入')}</button>
-                          <button
-                            class="remove-attachment"
-                            aria-label={tr('删除 {name}', { name: attachment.file_name })}
-                            disabled={attachmentBusy || workspace.request.status === 'completed' || workspace.request.status === 'cancelled'}
-                            onclick={() => removeAttachment(attachment)}
-                          >×</button>
-                        </div>
-                      </div>
-                    {/each}
-                  </div>
-                {:else}
-                  <p class="rail-empty">{tr('截图和导入的文件会直接进入正文，也会在这里留档。')}</p>
-                {/if}
-              </section>
-
-              <section class="rambelle-note" aria-label={tr('Rambelle 状态')}>
-                <div>
-                  <span>RAMBELLE · ONLINE</span>
-                  <p>
-                    {feedbackResult
-                      ? tr('记录在案了，长官。')
-                      : rambleEngaged
-                        ? tr('我正在整理这次 Ramble。')
-                        : tr('档案已就绪，随时可以继续。')}
-                  </p>
-                </div>
-                <img src={rambelleStatusPortrait} alt="Rambelle" />
-              </section>
-
-              <section class="delivery-card">
-                <div class="delivery-status">
-                  <span class:ready={canSubmit}></span>
-                  <div>
-                    <strong>{feedbackResult ? tr('反馈包已归档') : 'Feedback Package'}</strong>
-                    <small>{feedbackResult ? tr('Agent 已可读取不可变结果') : tr('正文保存后即可提交')}</small>
-                  </div>
-                </div>
-                {#if feedbackResult}
-                  <code>{desktopPath(feedbackResult.directory_path)}</code>
-                  <button class="package-button" onclick={openFeedbackPackage}>{tr('打开 Feedback Package')}</button>
-                {:else}
-                  <button class="primary-button wide-button" disabled={!canSubmit} onclick={submitFeedback}>
-                    {submitting ? tr('正在发布…') : tr('提交反馈')}
-                  </button>
-                {/if}
-              </section>
-            </aside>
-          </div>
-        </div>
-      {:else}
-        <div class="workspace-placeholder">
-          <span class="placeholder-mark">↙</span>
-          <strong>{tr('选择一个请求开始体验')}</strong>
-          <p>{tr('任务清单和你的 Markdown 草稿都会持久保存在本机。')}</p>
-        </div>
-      {/if}
-
-      {#if pageError}
-        <div class="error-banner" role="alert">
-          <strong>{tr('工作台暂时无法完成操作')}</strong>
-          <span>{pageError}</span>
-        </div>
-      {/if}
-    </section>
+    <WorkspacePanel
+      bind:this={workspacePanel}
+      bind:taskBriefOpen
+      {loadingWorkspace}
+      {workspace}
+      {feedbackResult}
+      {pageError}
+      {draftBody}
+      {savedRevision}
+      {savePhase}
+      {attachmentPreviews}
+      {dragActive}
+      {attachmentMessage}
+      {saveMessage}
+      {rambelleStatusPortrait}
+      {rambleEngaged}
+      {rambleActive}
+      {ramblePhase}
+      {rambleBusy}
+      {rambleStartedOnce}
+      {voiceDevice}
+      {voiceChunkIndex}
+      {voicePartial}
+      {voiceLevel}
+      {rambleMessage}
+      {attachmentBusy}
+      {canSubmit}
+      {submitting}
+      {adapterPresentation}
+      {formatTime}
+      onReload={() => void reloadWorkspace()}
+      onDraftChange={updateDraft}
+      onToggleRamble={() => void toggleRamble()}
+      onExitRamble={() => void exitRamble()}
+      onStartScreenCapture={() => void startScreenCapture()}
+      onImportClipboard={() => void importClipboardNow()}
+      onFileSelection={handleFileSelection}
+      onInsertAttachment={insertExistingAttachment}
+      onRemoveAttachment={(attachment) => void removeAttachment(attachment)}
+      onOpenPackage={() => void openFeedbackPackage()}
+      onSubmit={() => void submitFeedback()}
+    />
 
     {#if resumePrompt}
-      <div
-        class="resume-prompt-backdrop"
-        role="presentation"
-        onclick={(event) => {
-          if (event.target === event.currentTarget) dismissResumePrompt()
-        }}
-      >
-        <div
-          class="resume-prompt-dialog"
-          role="dialog"
-          aria-modal="true"
-          aria-labelledby="resume-prompt-title"
-        >
-          <div class="resume-prompt-header">
-            <span class="resume-prompt-kicker">WAKE · GENERIC</span>
-            <h2 id="resume-prompt-title">{resumePrompt.title}</h2>
-            <p>{resumePrompt.body}</p>
-          </div>
-          <div class="resume-prompt-meta">
-            <span>{tr('宿主')}</span>
-            <strong>{resumePrompt.host_label}</strong>
-            <span>request_id</span>
-            <code>{resumePrompt.request_id}</code>
-          </div>
-          <label class="resume-prompt-label" for="resume-prompt-text">{tr('恢复提示（复制到宿主对话）')}</label>
-          <textarea
-            id="resume-prompt-text"
-            class="resume-prompt-text"
-            readonly
-            rows="4"
-            value={resumePrompt.resume_prompt}
-          ></textarea>
-          <div class="resume-prompt-actions">
-            <button class="primary-button" onclick={copyResumePrompt}>
-              {resumeCopyState === 'copied'
-                ? tr('已复制')
-                : resumeCopyState === 'failed'
-                  ? tr('复制失败，请手动选择')
-                  : tr('复制恢复提示')}
-            </button>
-            <button class="secondary-button" onclick={dismissResumePrompt}>{tr('知道了')}</button>
-          </div>
-        </div>
-      </div>
+      <ResumePromptDialog
+        prompt={resumePrompt}
+        copyState={resumeCopyState}
+        onCopy={() => void copyResumePrompt()}
+        onDismiss={dismissResumePrompt}
+      />
     {/if}
   </div>
 </main>
