+++ apps/desktop/src/ScreenshotOverlay.svelte
@@ -1,30 +1,11 @@
 <script lang="ts">
   import { invoke } from '@tauri-apps/api/core'
   import { listen, type UnlistenFn } from '@tauri-apps/api/event'
-  import {
-    Check,
-    Circle,
-    Copy,
-    Ellipsis,
-    Grid3X3,
-    GripVertical,
-    Hash,
-    Highlighter,
-    Minus,
-    MousePointer2,
-    MoveUpRight,
-    Pencil,
-    RectangleHorizontal,
-    Redo2,
-    Trash2,
-    Type,
-    Undo2,
-    X,
-  } from '@lucide/svelte'
   import { onMount, tick } from 'svelte'
 
   import { t } from './lib/i18n'
   import { locale } from './lib/preferences'
+  import CaptureToolbar from './lib/screen-capture/CaptureToolbar.svelte'
   import {
     clampCaptureRectangle,
     distance,
@@ -959,90 +940,33 @@
   {/if}
 
   {#if selection}
-    <div
-      bind:clientWidth={toolbarWidth}
-      bind:clientHeight={toolbarHeight}
-      class="capture-toolbar"
-      class:popover-down={toolbarPopoverOpensDownward()}
-      data-capture-ui
-      style={captureToolbarStyle()}
-    >
-      <button
-        class="toolbar-drag"
-        aria-label="拖动工具栏"
-        title="拖动工具栏"
-        onmousedown={beginToolbarDrag}
-      ><GripVertical size={17} /></button>
-      <span class="divider"></span>
-      <div class="tool-group">
-        <button class:active={activeTool === 'select'} onclick={() => setTool('select')} title="选择/修改 · V"><MousePointer2 size={18} /></button>
-        <button class:active={activeTool === 'rectangle'} onclick={() => setTool('rectangle')} title="矩形 · R"><RectangleHorizontal size={18} /></button>
-        <button class:active={activeTool === 'ellipse'} onclick={() => setTool('ellipse')} title="圆形 · E"><Circle size={18} /></button>
-        <button class:active={activeTool === 'arrow'} onclick={() => setTool('arrow')} title="箭头 · A"><MoveUpRight size={18} /></button>
-        <button class:active={activeTool === 'pen'} onclick={() => setTool('pen')} title="画笔 · P"><Pencil size={18} /></button>
-        <button class:active={activeTool === 'text'} onclick={() => setTool('text')} title="文字 · T"><Type size={18} /></button>
-        <button class:active={activeTool === 'mosaic'} onclick={() => setTool('mosaic')} title="马赛克 · B"><Grid3X3 size={18} /></button>
-      </div>
-      <div class="popup-control">
-        <button
-          class="style-trigger"
-          class:active={stylePanelOpen}
-          aria-expanded={stylePanelOpen}
-          onclick={toggleStylePanel}
-          title="颜色与粗细"
-        ><i style={`--swatch:${currentColor};transform:scaleY(${currentStrokeWidth / 4})`}></i></button>
-        {#if stylePanelOpen}
-          <div class="toolbar-popover style-popover" aria-label="颜色与线条粗细">
-            <div class="palette" aria-label="标注颜色">
-              {#each colors as color}
-                <button
-                  class:active={currentColor === color}
-                  class="color-button"
-                  style={`--swatch:${color}`}
-                  onclick={() => setColor(color)}
-                  title={`颜色 ${color}`}
-                ></button>
-              {/each}
-            </div>
-            <span class="popover-divider"></span>
-            <div class="stroke-picker" aria-label="线条粗细">
-              {#each strokeWidths as width}
-                <button class:active={currentStrokeWidth === width} onclick={() => setStrokeWidth(width)} title={`${width}px`}>
-                  <i style={`height:${Math.max(2, width / 2)}px`}></i>
-                </button>
-              {/each}
-            </div>
-          </div>
-        {/if}
-      </div>
-      <div class="popup-control">
-        <button
-          class:active={overflowPanelOpen || activeTool === 'line' || activeTool === 'highlight' || activeTool === 'counter'}
-          aria-expanded={overflowPanelOpen}
-          onclick={toggleOverflowPanel}
-          title="更多工具"
-        ><Ellipsis size={18} /></button>
-        {#if overflowPanelOpen}
-          <div class="toolbar-popover more-popover" aria-label="更多工具">
-            <button class:active={activeTool === 'line'} onclick={() => setTool('line')} title="直线 · L"><Minus size={18} /></button>
-            <button class:active={activeTool === 'highlight'} onclick={() => setTool('highlight')} title="高亮 · H"><Highlighter size={18} /></button>
-            <button class:active={activeTool === 'counter'} onclick={() => setTool('counter')} title="序号 · N"><Hash size={18} /></button>
-            <button disabled={historyFuture.length === 0} onclick={redo} title="重做 · Ctrl/⌘ Shift Z"><Redo2 size={18} /></button>
-            <button disabled={!selectedAnnotationId} onclick={deleteSelected} title="删除选中标注 · Delete"><Trash2 size={18} /></button>
-          </div>
-        {/if}
-      </div>
-      <span class="divider"></span>
-      <div class="tool-group">
-        <button disabled={historyPast.length === 0} onclick={undo} title="撤销 · Ctrl/⌘ Z"><Undo2 size={18} /></button>
-      </div>
-      <span class="divider"></span>
-      <div class="tool-group actions">
-        <button onclick={() => finalize(true)} title="复制并插入"><Copy size={18} /></button>
-        <button class="confirm" onclick={() => finalize(false)} title="插入文档 · Enter"><Check size={19} /></button>
-        <button class="cancel" onclick={cancelCapture} title="取消 · Esc"><X size={19} /></button>
-      </div>
-    </div>
+    <CaptureToolbar
+      bind:toolbarWidth
+      bind:toolbarHeight
+      toolbarStyle={captureToolbarStyle()}
+      popoverDown={toolbarPopoverOpensDownward()}
+      {activeTool}
+      {stylePanelOpen}
+      {overflowPanelOpen}
+      {currentColor}
+      {currentStrokeWidth}
+      {colors}
+      {strokeWidths}
+      canUndo={historyPast.length > 0}
+      canRedo={historyFuture.length > 0}
+      canDelete={selectedAnnotationId !== null}
+      onBeginDrag={beginToolbarDrag}
+      onSetTool={setTool}
+      onToggleStylePanel={toggleStylePanel}
+      onToggleOverflowPanel={toggleOverflowPanel}
+      onSetColor={setColor}
+      onSetStrokeWidth={setStrokeWidth}
+      onUndo={undo}
+      onRedo={redo}
+      onDelete={deleteSelected}
+      onFinalize={(copyToClipboard) => void finalize(copyToClipboard)}
+      onCancel={() => void cancelCapture()}
+    />
   {/if}
 
   {#if errorMessage}
@@ -1057,217 +981,3 @@
     <div class="completing-mask" data-capture-ui><span>正在处理截图…</span></div>
   {/if}
 </main>
-
-<style>
-  .capture-editor {
-    position: fixed;
-    inset: 0;
-    overflow: hidden;
-    color: #f7fbff;
-    background: #000;
-    cursor: crosshair;
-    -webkit-user-select: none;
-    user-select: none;
-    touch-action: none;
-  }
-  .capture-editor * { -webkit-user-select: none; user-select: none; }
-  .capture-editor.has-selection { cursor: crosshair; }
-  .capture-image, .annotation-layer, .image-mask {
-    position: absolute;
-    display: block;
-  }
-  .capture-image { pointer-events: none; }
-  .annotation-layer { pointer-events: none; }
-  .image-mask { background: rgb(4 8 12 / 46%); pointer-events: none; }
-  .smart-target, .selection-frame, .annotation-selection { position: absolute; pointer-events: none; }
-  .smart-target {
-    z-index: 4;
-    border: 2px solid #61d79c;
-    background: rgb(86 216 151 / 7%);
-    box-shadow: 0 0 0 1px rgb(0 0 0 / 28%);
-  }
-  .smart-target span, .selection-size {
-    position: absolute;
-    top: -27px;
-    left: 0;
-    max-width: 260px;
-    overflow: hidden;
-    padding: 4px 8px;
-    border-radius: 6px;
-    color: white;
-    background: rgb(16 25 32 / 88%);
-    font-size: 10px;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-  }
-  .selection-frame {
-    z-index: 5;
-    border: 2px solid #61d79c;
-    box-shadow: 0 0 0 9999px rgb(4 8 12 / 46%);
-  }
-  .selection-size { top: auto; bottom: -27px; }
-  .resize-handle, .annotation-handle {
-    position: absolute;
-    width: 11px;
-    height: 11px;
-    padding: 0;
-    border: 2px solid #10261d;
-    border-radius: 50%;
-    background: #72e5ab;
-    pointer-events: auto;
-  }
-  .nw { top: -6px; left: -6px; cursor: nwse-resize; }
-  .ne { top: -6px; right: -6px; cursor: nesw-resize; }
-  .se { right: -6px; bottom: -6px; cursor: nwse-resize; }
-  .sw { bottom: -6px; left: -6px; cursor: nesw-resize; }
-  .annotation-selection {
-    z-index: 6;
-    border: 1px dashed #72c8ff;
-    background: rgb(74 164 231 / 5%);
-  }
-  .annotation-handle { width: 9px; height: 9px; border-color: #153348; background: #72c8ff; }
-  .text-editor {
-    position: absolute;
-    z-index: 12;
-    width: min(320px, 35vw);
-    min-width: 100px;
-    min-height: 38px;
-    padding: 5px 7px;
-    border: 1px dashed currentColor;
-    outline: none;
-    background: rgb(10 14 19 / 48%);
-    font-weight: 600;
-    line-height: 1.25;
-    resize: both;
-    -webkit-user-select: text;
-    user-select: text;
-  }
-  .capture-help, .capture-status {
-    position: fixed;
-    top: 22px;
-    left: 50%;
-    z-index: 20;
-    display: grid;
-    gap: 3px;
-    justify-items: center;
-    padding: 10px 17px;
-    border: 1px solid rgb(255 255 255 / 18%);
-    border-radius: 11px;
-    background: rgb(16 23 31 / 84%);
-    box-shadow: 0 12px 36px rgb(0 0 0 / 26%);
-    backdrop-filter: blur(12px);
-    transform: translateX(-50%);
-  }
-  .capture-help strong, .capture-status strong { font-size: 12px; }
-  .capture-help span { color: rgb(255 255 255 / 64%); font-size: 9px; }
-  .capture-toolbar {
-    position: fixed;
-    z-index: 30;
-    display: flex;
-    max-width: calc(100vw - 28px);
-    align-items: center;
-    gap: 3px;
-    padding: 5px;
-    overflow: visible;
-    border: 1px solid rgb(255 255 255 / 15%);
-    border-radius: 11px;
-    background: rgb(18 25 34 / 94%);
-    box-shadow: 0 16px 50px rgb(0 0 0 / 38%);
-    backdrop-filter: blur(15px);
-  }
-  .tool-group, .palette, .stroke-picker { display: flex; flex: 0 0 auto; gap: 3px; align-items: center; }
-  .capture-toolbar button {
-    display: grid;
-    width: 30px;
-    height: 30px;
-    flex: 0 0 auto;
-    place-items: center;
-    padding: 0;
-    border: 1px solid transparent;
-    border-radius: 7px;
-    color: rgb(239 247 255 / 76%);
-    background: transparent;
-    cursor: pointer;
-  }
-  .capture-toolbar button:hover:not(:disabled) { color: white; background: rgb(255 255 255 / 9%); }
-  .capture-toolbar button.active { border-color: rgb(93 199 255 / 36%); color: #8bd5ff; background: rgb(62 159 220 / 18%); }
-  .capture-toolbar button:disabled { cursor: default; opacity: 0.3; }
-  .capture-toolbar .toolbar-drag { width: 24px; color: rgb(239 247 255 / 42%); cursor: grab; touch-action: none; }
-  .capture-toolbar .toolbar-drag:active { cursor: grabbing; }
-  .divider { width: 1px; height: 22px; flex: 0 0 auto; margin: 0 2px; background: rgb(255 255 255 / 12%); }
-  .popup-control { position: relative; display: flex; flex: 0 0 auto; }
-  .toolbar-popover {
-    position: absolute;
-    right: 0;
-    bottom: calc(100% + 8px);
-    z-index: 45;
-    display: flex;
-    align-items: center;
-    gap: 5px;
-    padding: 6px;
-    border: 1px solid rgb(255 255 255 / 15%);
-    border-radius: 10px;
-    background: rgb(18 25 34 / 97%);
-    box-shadow: 0 12px 36px rgb(0 0 0 / 42%);
-    backdrop-filter: blur(15px);
-  }
-  .capture-toolbar.popover-down .toolbar-popover { top: calc(100% + 8px); bottom: auto; }
-  .style-popover { right: -34px; }
-  .more-popover {
-    right: -6px;
-    display: grid;
-    grid-template-columns: repeat(3, 30px);
-    gap: 3px;
-  }
-  .capture-toolbar .style-trigger i {
-    display: block;
-    width: 17px;
-    height: 4px;
-    border: 1px solid rgb(255 255 255 / 45%);
-    border-radius: 6px;
-    background: var(--swatch);
-  }
-  .popover-divider { width: 1px; height: 24px; flex: 0 0 auto; background: rgb(255 255 255 / 12%); }
-  .palette { gap: 5px; }
-  .capture-toolbar .color-button {
-    width: 20px;
-    height: 20px;
-    border: 2px solid rgb(255 255 255 / 22%);
-    border-radius: 50%;
-    background: var(--swatch);
-  }
-  .capture-toolbar .color-button.active { border-color: white; box-shadow: 0 0 0 2px rgb(68 171 239 / 70%); }
-  .stroke-picker button { width: 25px; height: 30px; }
-  .stroke-picker i { display: block; width: 15px; border-radius: 9px; background: currentColor; }
-  .actions .confirm { color: #edfff5; background: rgb(49 166 101 / 54%); }
-  .actions .cancel { color: #ffafb5; }
-  .capture-error {
-    position: fixed;
-    top: 50%;
-    left: 50%;
-    z-index: 60;
-    display: flex;
-    width: min(430px, calc(100vw - 42px));
-    flex-direction: column;
-    gap: 9px;
-    padding: 18px;
-    border-radius: 13px;
-    color: #531f24;
-    background: #fff;
-    box-shadow: 0 22px 70px rgb(0 0 0 / 46%);
-    transform: translate(-50%, -50%);
-  }
-  .capture-error strong { font-size: 13px; }
-  .capture-error span { font-size: 10px; line-height: 1.5; }
-  .capture-error button { align-self: flex-end; padding: 6px 13px; border: 0; border-radius: 7px; color: white; background: #294e39; cursor: pointer; }
-  .completing-mask {
-    position: fixed;
-    inset: 0;
-    z-index: 80;
-    display: grid;
-    place-items: center;
-    background: rgb(7 10 14 / 28%);
-    cursor: wait;
-  }
-  .completing-mask span { padding: 10px 16px; border-radius: 9px; background: rgb(18 25 34 / 88%); font-size: 11px; }
-</style>
