+++ apps/desktop/src/styles/workbench-document.css
@@ -1,0 +1,233 @@
+.task-sheet {
+  overflow: hidden;
+  border-color: color-mix(in srgb, var(--blue) 22%, var(--line-soft));
+  box-shadow: 0 12px 32px rgb(42 70 101 / 7%);
+}
+
+.task-sheet-toggle {
+  display: flex;
+  width: 100%;
+  align-items: center;
+  justify-content: space-between;
+  gap: 16px;
+  padding: 12px 14px 12px 13px;
+  border: 0;
+  color: var(--ink);
+  background: transparent;
+  cursor: pointer;
+}
+
+.task-sheet-toggle:hover {
+  background: color-mix(in srgb, var(--blue-soft) 48%, transparent);
+}
+
+.task-sheet-toggle > span {
+  display: flex;
+  min-width: 0;
+  align-items: center;
+  gap: 10px;
+}
+
+.task-sheet-toggle i {
+  display: grid;
+  width: 28px;
+  height: 28px;
+  place-items: center;
+  border: 1px solid color-mix(in srgb, var(--blue) 28%, var(--line-soft));
+  border-radius: 8px;
+  color: var(--blue-strong);
+  background: color-mix(in srgb, var(--blue-soft) 82%, var(--surface-raised));
+  font-size: 9px;
+  font-style: normal;
+  font-weight: 760;
+}
+
+.task-sheet-toggle strong {
+  font-size: 12px;
+  letter-spacing: -0.01em;
+}
+
+.task-sheet-toggle em {
+  color: var(--ink-faint);
+  font-size: 9px;
+  font-style: normal;
+  font-weight: 520;
+}
+
+.task-sheet-toggle-icon {
+  display: grid;
+  width: 30px;
+  height: 30px;
+  flex: 0 0 auto;
+  place-items: center;
+  border: 1px solid var(--line-soft);
+  border-radius: 9px;
+  color: var(--ink-soft);
+  background: color-mix(in srgb, var(--surface-raised) 76%, transparent);
+  transition: color 140ms ease, background 140ms ease;
+}
+
+.task-sheet-toggle:hover .task-sheet-toggle-icon {
+  color: var(--blue-strong);
+  background: var(--blue-soft);
+}
+
+.task-sheet-toggle-icon svg {
+  transition: transform 160ms ease;
+}
+
+.task-sheet-toggle-icon.open svg {
+  transform: rotate(180deg);
+}
+
+.task-sheet-body {
+  display: grid;
+  grid-template-columns: minmax(0, 1fr);
+  gap: 10px;
+  padding: 11px;
+  border-top: 1px solid var(--line-soft);
+  background: color-mix(in srgb, var(--surface-tint) 68%, var(--surface-raised));
+}
+
+.brief-section {
+  min-width: 0;
+  padding: 13px 14px;
+  border: 1px solid var(--line-soft);
+  border-radius: 10px;
+  background: var(--surface-raised);
+}
+
+.brief-summary-section {
+  border-color: color-mix(in srgb, var(--blue) 22%, var(--line-soft));
+  background:
+    linear-gradient(135deg, color-mix(in srgb, var(--blue-soft) 72%, transparent), transparent 72%),
+    var(--surface-raised);
+}
+
+.brief-summary {
+  margin: 0;
+  color: var(--ink);
+  font-size: 12px;
+  font-weight: 620;
+  line-height: 1.55;
+}
+
+.actions {
+  display: grid;
+  gap: 9px;
+  margin: 0;
+  padding: 0;
+  list-style: none;
+  counter-reset: action;
+}
+
+.actions li {
+  display: grid;
+  grid-template-columns: 23px 1fr;
+  gap: 9px;
+  align-items: start;
+  counter-increment: action;
+}
+
+.actions li > span {
+  display: grid;
+  width: 21px;
+  height: 21px;
+  place-items: center;
+  overflow: hidden;
+  border: 1px solid #bdd4eb;
+  border-radius: 50%;
+  color: transparent;
+  font-size: 0;
+}
+
+.actions li > span::before {
+  color: var(--blue-strong);
+  font-size: 9px;
+  font-weight: 700;
+  content: counter(action);
+}
+
+.actions p {
+  margin: 2px 0 0;
+  color: #3e5065;
+  font-size: 11px;
+  line-height: 1.48;
+}
+
+.editor-section {
+  position: relative;
+  min-width: 0;
+  padding: 14px;
+}
+
+.editor-section.drag-active::after {
+  position: absolute;
+  inset: 8px;
+  display: grid;
+  place-items: center;
+  border: 2px dashed var(--cyan);
+  border-radius: 10px;
+  color: #157f79;
+  background: rgb(235 249 247 / 95%);
+  font-size: 16px;
+  font-weight: 680;
+  content: "松开以导入文档流";
+  pointer-events: none;
+  z-index: 8;
+}
+
+.editor-heading {
+  justify-content: space-between;
+  gap: 16px;
+  padding: 1px 2px 10px;
+}
+
+.editor-heading h3 {
+  margin: 0;
+  color: #213650;
+  font-size: 17px;
+  font-weight: 670;
+  letter-spacing: -0.02em;
+}
+
+.save-state {
+  display: flex;
+  align-items: center;
+  gap: 6px;
+  color: #41847f;
+  font-size: 9px;
+  font-weight: 620;
+}
+
+.save-state.failed,
+.inline-error {
+  color: var(--danger);
+}
+
+.save-dot {
+  width: 7px;
+  height: 7px;
+  border-radius: 50%;
+  background: var(--cyan);
+  box-shadow: 0 0 0 3px rgb(87 198 192 / 12%);
+}
+
+.save-state.failed .save-dot {
+  background: var(--danger);
+}
+
+.editor-footer {
+  justify-content: space-between;
+  gap: 12px;
+  padding: 10px 3px 0;
+  color: var(--ink-faint);
+  font-size: 8px;
+}
+
+.inline-error {
+  margin: 9px 2px 0;
+  font-size: 10px;
+  line-height: 1.45;
+}
+
