+++ apps/desktop/src/styles/workbench-shell.css
@@ -1,0 +1,427 @@
+.shell {
+  width: 100%;
+  height: 100%;
+  overflow: hidden;
+  border: 1px solid color-mix(in srgb, var(--line) 86%, transparent);
+  border-radius: 16px;
+  background: var(--surface);
+}
+
+.workbench {
+  display: grid;
+  grid-template-columns: 282px minmax(0, 1fr);
+  height: calc(100% - 46px);
+  min-height: 0;
+}
+
+.inbox-panel {
+  display: flex;
+  min-width: 0;
+  min-height: 0;
+  flex-direction: column;
+  padding: 18px 12px 12px;
+  border-right: 1px solid var(--line-soft);
+  background: rgb(247 249 252 / 86%);
+  backdrop-filter: blur(16px);
+}
+
+.panel-heading,
+.workspace-heading,
+.editor-heading,
+.editor-footer,
+.workspace-meta,
+.request-meta,
+.rail-heading,
+.delivery-status {
+  display: flex;
+  align-items: center;
+}
+
+.panel-heading {
+  justify-content: space-between;
+  padding: 2px 8px 12px;
+}
+
+.panel-heading h1 {
+  margin: 0;
+  color: var(--ink);
+  font-size: 20px;
+  font-weight: 680;
+  letter-spacing: -0.025em;
+}
+
+.eyebrow {
+  margin: 0 0 5px;
+  color: #5e87b3;
+  font-size: 9px;
+  font-weight: 760;
+  letter-spacing: 0.16em;
+}
+
+h1,
+h2,
+h3,
+p {
+  margin-top: 0;
+}
+
+h1,
+h2,
+h3 {
+  font-family: "Segoe UI Variable", "Segoe UI", "Microsoft YaHei UI", sans-serif;
+}
+
+.icon-button,
+.secondary-button,
+.package-button {
+  border: 1px solid var(--line);
+  color: #526a84;
+  background: var(--surface-raised);
+  cursor: pointer;
+}
+
+.icon-button {
+  display: grid;
+  width: 30px;
+  height: 30px;
+  place-items: center;
+  padding: 0;
+  border-radius: 8px;
+  font-size: 15px;
+}
+
+.icon-button:hover,
+.secondary-button:hover,
+.package-button:hover {
+  border-color: #88acd1;
+  color: var(--blue-strong);
+  background: #f5faff;
+}
+
+.inbox-tabs {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 3px;
+  margin: 0 6px 12px;
+  padding: 3px;
+  border: 1px solid var(--line-soft);
+  border-radius: 9px;
+  background: #edf2f7;
+}
+
+.inbox-tabs button {
+  padding: 7px;
+  border: 0;
+  border-radius: 6px;
+  color: #718296;
+  background: transparent;
+  cursor: pointer;
+  font-size: 10px;
+  font-weight: 620;
+}
+
+.inbox-tabs button.active {
+  color: var(--blue-strong);
+  background: #fff;
+  box-shadow: 0 2px 9px rgb(42 81 122 / 8%);
+}
+
+.inbox-panel nav {
+  display: grid;
+  gap: 6px;
+  min-height: 0;
+  overflow: auto;
+  padding: 0 4px 10px;
+}
+
+.request-card {
+  position: relative;
+  width: 100%;
+  padding: 12px 12px 11px 14px;
+  overflow: hidden;
+  border: 1px solid transparent;
+  border-radius: 10px;
+  color: var(--ink);
+  text-align: left;
+  background: transparent;
+  cursor: pointer;
+}
+
+.request-card::before {
+  position: absolute;
+  inset: 11px auto 11px 0;
+  width: 2px;
+  border-radius: 2px;
+  background: transparent;
+  content: "";
+}
+
+.request-card:hover {
+  background: rgb(255 255 255 / 64%);
+}
+
+.request-card.active {
+  border-color: #a9c6e5;
+  background: #fff;
+  box-shadow: 0 10px 26px rgb(47 91 137 / 8%);
+}
+
+.request-card.active::before {
+  background: var(--blue-strong);
+}
+
+.request-meta {
+  justify-content: space-between;
+  gap: 8px;
+  margin-bottom: 7px;
+}
+
+.request-meta b {
+  overflow: hidden;
+  color: #4b80b8;
+  font-size: 9px;
+  font-weight: 670;
+  letter-spacing: 0.02em;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.request-meta em {
+  flex: 0 0 auto;
+  padding: 2px 5px;
+  border-radius: 5px;
+  color: #168a83;
+  background: var(--cyan-soft);
+  font-size: 8px;
+  font-style: normal;
+}
+
+.request-card > strong {
+  display: -webkit-box;
+  overflow: hidden;
+  margin-bottom: 8px;
+  font-size: 12px;
+  font-weight: 650;
+  line-height: 1.4;
+  -webkit-box-orient: vertical;
+  -webkit-line-clamp: 2;
+}
+
+.request-card small {
+  color: var(--ink-faint);
+  font-size: 9px;
+}
+
+.request-byline,
+.adapter-chip {
+  display: flex;
+  align-items: center;
+  gap: 5px;
+}
+
+.adapter-mark {
+  display: inline-grid;
+  width: 16px;
+  height: 16px;
+  flex: 0 0 auto;
+  place-items: center;
+}
+
+.adapter-mark svg {
+  width: 100%;
+  height: 100%;
+}
+
+.adapter-mark svg[role="img"] {
+  fill: currentColor;
+}
+
+.empty-state {
+  display: grid;
+  gap: 7px;
+  padding: 24px 14px;
+  color: var(--ink-faint);
+  font-size: 11px;
+  line-height: 1.55;
+}
+
+.empty-state strong {
+  color: var(--ink);
+  font-size: 13px;
+}
+
+.connection-card {
+  position: relative;
+  display: block;
+  width: 100%;
+  margin-top: auto;
+  padding: 10px;
+  border: 1px solid var(--line-soft);
+  border-radius: 9px;
+  color: inherit;
+  background: rgb(255 255 255 / 54%);
+  text-align: left;
+  cursor: pointer;
+  transition: border-color 150ms ease, background 150ms ease, transform 150ms ease;
+}
+
+.connection-card:hover {
+  border-color: color-mix(in srgb, var(--blue) 50%, var(--line-soft));
+  background: color-mix(in srgb, var(--surface-raised) 88%, var(--blue-soft));
+  transform: translateY(-1px);
+}
+
+.connection-card svg {
+  position: absolute;
+  top: 10px;
+  right: 10px;
+  color: var(--blue-strong);
+}
+
+.connection-card span,
+.connection-card code {
+  display: block;
+}
+
+.connection-card span {
+  margin-bottom: 4px;
+  color: var(--ink-faint);
+  font-size: 8px;
+  font-weight: 700;
+  letter-spacing: 0.12em;
+  text-transform: uppercase;
+}
+
+.connection-card code,
+.context code,
+.delivery-card code {
+  overflow-wrap: anywhere;
+  color: #5c7188;
+  font-family: ui-monospace, SFMono-Regular, Menlo, Consolas, monospace;
+  font-size: 9px;
+}
+
+.workspace-panel {
+  position: relative;
+  min-width: 0;
+  min-height: 0;
+  overflow-x: hidden;
+  overflow-y: auto;
+  padding: 22px clamp(44px, 4vw, 62px) 34px clamp(20px, 2.6vw, 38px);
+  background-image:
+    linear-gradient(rgb(237 242 247 / 84%), rgb(237 242 247 / 84%)),
+    url("../assets/rambelle-vault-pattern.png");
+  background-repeat: no-repeat, repeat;
+  background-size: cover, 408px 408px;
+  scrollbar-gutter: stable both-edges;
+}
+
+.workspace-stage {
+  width: min(1500px, 100%);
+  min-width: 0;
+  margin: 0 auto;
+}
+
+.workspace-heading {
+  align-items: flex-start;
+  justify-content: space-between;
+  gap: 18px;
+  margin-bottom: 15px;
+  padding: 0 2px;
+}
+
+.workspace-heading-copy {
+  min-width: 0;
+}
+
+.workspace-heading h2 {
+  overflow: hidden;
+  margin: 7px 0 5px;
+  color: var(--ink);
+  font-size: clamp(20px, 2vw, 27px);
+  font-weight: 670;
+  letter-spacing: -0.03em;
+  line-height: 1.16;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.workspace-heading p {
+  margin-bottom: 0;
+  color: var(--ink-faint);
+  font-size: 9px;
+}
+
+.workspace-meta {
+  flex-wrap: wrap;
+  gap: 5px;
+}
+
+.workspace-meta span {
+  padding: 3px 6px;
+  border: 1px solid #d5e1ed;
+  border-radius: 5px;
+  color: #57728f;
+  background: rgb(247 250 253 / 75%);
+  font-size: 8px;
+  font-weight: 620;
+}
+
+.workspace-meta span.status-chip {
+  border-color: #bfe5df;
+  color: #168a83;
+  background: var(--cyan-soft);
+}
+
+.workspace-meta span.adapter-chip {
+  display: flex;
+  align-items: center;
+  gap: 4px;
+  padding-block: 2px;
+}
+
+.workspace-meta .adapter-mark {
+  width: 14px;
+  height: 14px;
+}
+
+.secondary-button,
+.package-button {
+  flex: 0 0 auto;
+  padding: 7px 10px;
+  border-radius: 8px;
+  font-size: 10px;
+  font-weight: 620;
+}
+
+.compact-button {
+  margin-top: 2px;
+}
+
+.workspace-columns {
+  display: grid;
+  min-width: 0;
+  grid-template-columns: minmax(0, 1fr) minmax(250px, 292px);
+  gap: 14px;
+  align-items: start;
+}
+
+.document-column {
+  display: grid;
+  min-width: 0;
+  gap: 10px;
+}
+
+.task-sheet,
+.editor-section,
+.ramble-console,
+.tool-card,
+.attachments-card,
+.rambelle-note,
+.delivery-card {
+  border: 1px solid var(--line-soft);
+  border-radius: 12px;
+  background: rgb(255 255 255 / 90%);
+  box-shadow: var(--shadow);
+}
+
